Handle direction when moving to connected road
Previously the SetLanePos function tried to deduce direction based on
previous and current lane indices after moving to a connected road.
However, to cover all cases, it's better to resolve direction in the
MoveToConnectingRoad where the contact points are known.

Please enter the commit message for your changes. Lines starting